TL Modify

+++ GitProject01/src/main/java/com/nt/service/ArithemeticOperations.java
@@ -2,7 +2,7 @@ package com.nt.service;
 
 public class ArithemeticOperations {
 	
-	public double sum(double x, double y) {
+	public int  sum(int x, int y) {
 		return x+y;
 	}
 }
